CASMPET-5155: Calculate certificate name
The hardcoded certificate name made it such that there could only
be 1 instance of the cray-oauth2-proxy chart deployed. This change
makes it so that multiple instances of the chart can be deployed.

This chart hasn't been included in the CSM manifest so I included
the version change in this change.