CASMPET-5155: Chart to create separate oauth2-proxy instances for Bi-CAN
This change adds a new chart that deploys multiple instances of the
cray-oauth2-proxy chart.

NOTE: You might wonder why this can't be done in the loftsman
manifest using unique `releaseName` values. I tried this and it
didn't work. The other tools (e.g., manifestgen) ignore that field
and only use the name, so this loftsman feature is not usable.

With bifurcated CAN we need oauth2-proxy instances listening on
different interfaces. This is done by setting the
`metallb.universe.tf/address-pool` on the services
(metalLbAddressPool in values.yaml). Also the oauth2-proxy
instances are proxying different sets of apps on different
hostnames.

keycloak-setup will create the oauth2ClientSecrets.

The oauth2ProxySecrets will be created using Sealed Secrets, this
is in the customizations.yaml.

The hosts for each cray-oauth2-proxy instance will be set in
customizations.yaml.